devtools/client/webconsole/new-console-output/test/components/test_message-container.html

<!DOCTYPE HTML>
<html lang="en">
<head>
  <meta charset="utf8">
  <title>Test for MessageContainer component</title>
  <script type="text/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript;version=1.8" src="head.js"></script>
  <!-- Any copyright is dedicated to the Public Domain.
     - http://creativecommons.org/publicdomain/zero/1.0/ -->
</head>
<body>
<p>Test for MessageContainer component</p>

<script type="text/javascript;version=1.8">
window.onload = Task.async(function* () {
  const React = require("devtools/client/shared/vendor/react");
  const TestUtils = React.addons.TestUtils;
  const { prepareMessage } = require("devtools/client/webconsole/new-console-output/actions/messages");

  const { MessageContainer } = require("devtools/client/webconsole/new-console-output/components/message-container");
  const { ConsoleApiCall } = require("devtools/client/webconsole/new-console-output/components/message-types/console-api-call");

  const packet = yield getPacket("console.log('foobar', 'test')", "consoleAPICall");
  const message = prepareMessage(packet);
  const el = React.createElement(MessageContainer, { message });
  const renderer = TestUtils.createRenderer();
  renderer.render(el, {});
  const result = renderer.getRenderOutput();

  is(result.type, ConsoleApiCall,
    "MessageContainer nests the correct component based on packet");

  SimpleTest.finish()
});
</script>
</body>
</html>
